# internal/services/gateway/docs/swagger.yaml
definitions:
...
  models.AgentLifecycleStatus:
    enum:
    - idle
    - queued
    - running
    - waiting
    - completed
    - failed
    - offline
    type: string
    x-enum-varnames:
    - AgentLifecycleStatusIdle
    - AgentLifecycleStatusQueued
    - AgentLifecycleStatusRunning
    - AgentLifecycleStatusWaiting
    - AgentLifecycleStatusCompleted
    - AgentLifecycleStatusFailed
    - AgentLifecycleStatusOffline
...
  models.ModelProvenanceResponse:
    properties:
      window:
        items:
          type: integer
        type: array
    type: object
...
  models.ObserveProducerAgentStateRequest:
    properties:
      agent_id:
        type: string
      cli_session_id:
        type: string
      display_name:
        type: string
      model:
        type: string
      observed_at:
        type: string
      role:
        type: string
      run_id:
        type: string
      schema_version:
        type: string
      status:
        $ref: '#/definitions/models.AgentLifecycleStatus'
      task_id:
        type: string
      web_session_id:
        type: string
    type: object
  models.ObserveProducerRunStateRequest:
    properties:
      active_task_id:
        type: string
      cli_session_id:
        type: string
      completed_tasks:
        type: integer
      display_name:
        type: string
      ended_at:
        type: string
      observed_at:
        type: string
      run_id:
        type: string
      run_kind:
        $ref: '#/definitions/models.RunKind'
      schema_version:
        type: string
      started_at:
        type: string
      status:
        $ref: '#/definitions/models.RunLifecycleStatus'
      total_tasks:
        type: integer
      web_session_id:
        type: string
    type: object
...
  models.RunKind:
    enum:
    - investigation
    - eval
    - demo
    - workflow
    type: string
    x-enum-varnames:
    - RunKindInvestigation
    - RunKindEval
    - RunKindDemo
    - RunKindWorkflow
  models.RunLifecycleStatus:
    enum:
    - queued
    - running
    - waiting
    - completed
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - RunLifecycleStatusQueued
    - RunLifecycleStatusRunning
    - RunLifecycleStatusWaiting
    - RunLifecycleStatusCompleted
    - RunLifecycleStatusFailed
    - RunLifecycleStatusCancelled
...
info:
  contact: {}
paths:
...
  /api/v1/inference/model-provenance/attestations/{provider_attempt_id}:
    get:
      description: Returns the canonical model provenance attestation window for one
        provider attempt ID (mTLS owner CLI only).
      parameters:
      - description: Provider attempt ID
        in: path
        name: provider_attempt_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModelProvenanceResponse'
        "404":
          description: Model provenance evidence not found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "503":
          description: Model provenance store unavailable
          schema:
            type: string
      summary: Get model provenance attestation evidence
      tags:
      - inference
...
  /api/v1/observe/producer/agent-state:
    post:
      consumes:
      - application/json
      description: Accepts a typed agent state update from an mTLS-authenticated app
        workload (the g8ee ensemble). The gateway derives user_id from the peer certificate,
        persists the agent state projection, and emits an app.agent.status.updated
        SSE event after successful persistence.
      parameters:
      - description: Agent state update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ObserveProducerAgentStateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: accepted
          schema:
            type: string
        "400":
          description: Invalid transition, stale update, or missing fields
          schema:
            type: string
        "401":
          description: Unauthorized — mTLS user identity required
          schema:
            type: string
        "403":
          description: Forbidden — not an app workload or cross-user
          schema:
            type: string
        "500":
          description: Persistence failure
          schema:
            type: string
      summary: Push observe agent state
      tags:
      - observe
  /api/v1/observe/producer/run-state:
    post:
      consumes:
      - application/json
      description: Accepts a typed run state update from an mTLS-authenticated app
        workload (the g8ee ensemble). The gateway derives user_id from the peer certificate,
        persists the run state projection, and emits an app.run.status.updated SSE
        event after successful persistence.
      parameters:
      - description: Run state update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ObserveProducerRunStateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: accepted
          schema:
            type: string
        "400":
          description: Invalid transition, stale update, or missing fields
          schema:
            type: string
        "401":
          description: Unauthorized — mTLS user identity required
          schema:
            type: string
        "403":
          description: Forbidden — not an app workload or cross-user
          schema:
            type: string
        "500":
          description: Persistence failure
          schema:
            type: string
      summary: Push observe run state
      tags:
      - observe
...
swagger: "2.0"